refactor(generator): extract DbContextFinder from SymbolDiscovery
