+++ framework/SimpleModule.Generator/Discovery/SymbolDiscovery.cs
@@ -171,8 +171,13 @@ internal static class SymbolDiscovery
             // Collect unmatched items from this assembly
             var rawDbContexts = new List<DbContextInfo>();
             var rawEntityConfigs = new List<EntityConfigInfo>();
-            FindDbContextTypes(assembly.GlobalNamespace, "", rawDbContexts, cancellationToken);
-            FindEntityConfigTypes(
+            DbContextFinder.FindDbContextTypes(
+                assembly.GlobalNamespace,
+                "",
+                rawDbContexts,
+                cancellationToken
+            );
+            DbContextFinder.FindEntityConfigTypes(
                 assembly.GlobalNamespace,
                 "",
                 rawEntityConfigs,
@@ -758,211 +763,6 @@ internal static class SymbolDiscovery
         return 1; // Default: Scoped
     }
 
-    private static bool HasDbContextConstructorParam(INamedTypeSymbol typeSymbol)
-    {
-        foreach (var ctor in typeSymbol.Constructors)
-        {
-            if (ctor.DeclaredAccessibility != Accessibility.Public || ctor.IsStatic)
-                continue;
-
-            foreach (var param in ctor.Parameters)
-            {
-                var paramType = param.Type;
-                // Walk the base type chain to check for DbContext ancestry
-                var current = paramType.BaseType;
-                while (current != null)
-                {
-                    var baseFqn = current.ToDisplayString(SymbolDisplayFormat.FullyQualifiedFormat);
-                    if (
-                        baseFqn == "global::Microsoft.EntityFrameworkCore.DbContext"
-                        || baseFqn.StartsWith(
-                            "global::Microsoft.AspNetCore.Identity.EntityFrameworkCore.IdentityDbContext",
-                            StringComparison.Ordinal
-                        )
-                    )
-                    {
-                        return true;
-                    }
-
-                    current = current.BaseType;
-                }
-            }
-        }
-
-        return false;
-    }
-
-    private static void FindDbContextTypes(
-        INamespaceSymbol namespaceSymbol,
-        string moduleName,
-        List<DbContextInfo> dbContexts,
-        CancellationToken cancellationToken
-    )
-    {
-        foreach (var member in namespaceSymbol.GetMembers())
-        {
-            cancellationToken.ThrowIfCancellationRequested();
-
-            if (member is INamespaceSymbol childNamespace)
-            {
-                FindDbContextTypes(childNamespace, moduleName, dbContexts, cancellationToken);
-            }
-            else if (
-                member is INamedTypeSymbol typeSymbol
-                && !typeSymbol.IsAbstract
-                && !typeSymbol.IsStatic
-            )
-            {
-                // Walk base type chain looking for DbContext
-                var isDbContext = false;
-                var isIdentity = false;
-                string identityUserFqn = "";
-                string identityRoleFqn = "";
-                string identityKeyFqn = "";
-
-                var current = typeSymbol.BaseType;
-                while (current is not null)
-                {
-                    var baseFqn = current.OriginalDefinition.ToDisplayString(
-                        SymbolDisplayFormat.FullyQualifiedFormat
-                    );
-
-                    if (
-                        baseFqn
-                        == "global::Microsoft.AspNetCore.Identity.EntityFrameworkCore.IdentityDbContext<TUser, TRole, TKey>"
-                    )
-                    {
-                        isDbContext = true;
-                        isIdentity = true;
-                        if (current.TypeArguments.Length >= 3)
-                        {
-                            identityUserFqn = current
-                                .TypeArguments[0]
-                                .ToDisplayString(SymbolDisplayFormat.FullyQualifiedFormat);
-                            identityRoleFqn = current
-                                .TypeArguments[1]
-                                .ToDisplayString(SymbolDisplayFormat.FullyQualifiedFormat);
-                            identityKeyFqn = current
-                                .TypeArguments[2]
-                                .ToDisplayString(SymbolDisplayFormat.FullyQualifiedFormat);
-                        }
-                        break;
-                    }
-
-                    if (baseFqn == "global::Microsoft.EntityFrameworkCore.DbContext")
-                    {
-                        isDbContext = true;
-                        break;
-                    }
-
-                    current = current.BaseType;
-                }
-
-                if (!isDbContext)
-                    continue;
-
-                var info = new DbContextInfo
-                {
-                    FullyQualifiedName = typeSymbol.ToDisplayString(
-                        SymbolDisplayFormat.FullyQualifiedFormat
-                    ),
-                    ModuleName = moduleName,
-                    IsIdentityDbContext = isIdentity,
-                    IdentityUserTypeFqn = identityUserFqn,
-                    IdentityRoleTypeFqn = identityRoleFqn,
-                    IdentityKeyTypeFqn = identityKeyFqn,
-                    Location = SymbolHelpers.GetSourceLocation(typeSymbol),
-                };
-
-                // Collect DbSet<T> properties
-                foreach (var m in typeSymbol.GetMembers())
-                {
-                    if (
-                        m is IPropertySymbol prop
-                        && prop.DeclaredAccessibility == Accessibility.Public
-                        && !prop.IsStatic
-                        && prop.Type is INamedTypeSymbol propType
-                        && propType.IsGenericType
-                        && propType.OriginalDefinition.ToDisplayString(
-                            SymbolDisplayFormat.FullyQualifiedFormat
-                        ) == "global::Microsoft.EntityFrameworkCore.DbSet<TEntity>"
-                    )
-                    {
-                        var entityType = propType.TypeArguments[0];
-                        var entityFqn = entityType.ToDisplayString(
-                            SymbolDisplayFormat.FullyQualifiedFormat
-                        );
-                        var entityAssemblyName =
-                            entityType.ContainingAssembly?.Name ?? string.Empty;
-                        info.DbSets.Add(
-                            new DbSetInfo
-                            {
-                                PropertyName = prop.Name,
-                                EntityFqn = entityFqn,
-                                EntityAssemblyName = entityAssemblyName,
-                                EntityLocation = SymbolHelpers.GetSourceLocation(entityType),
-                            }
-                        );
-                    }
-                }
-
-                dbContexts.Add(info);
-            }
-        }
-    }
-
-    private static void FindEntityConfigTypes(
-        INamespaceSymbol namespaceSymbol,
-        string moduleName,
-        List<EntityConfigInfo> entityConfigs,
-        CancellationToken cancellationToken
-    )
-    {
-        foreach (var member in namespaceSymbol.GetMembers())
-        {
-            cancellationToken.ThrowIfCancellationRequested();
-
-            if (member is INamespaceSymbol childNamespace)
-            {
-                FindEntityConfigTypes(childNamespace, moduleName, entityConfigs, cancellationToken);
-            }
-            else if (
-                member is INamedTypeSymbol typeSymbol
-                && !typeSymbol.IsAbstract
-                && !typeSymbol.IsStatic
-            )
-            {
-                foreach (var iface in typeSymbol.AllInterfaces)
-                {
-                    if (
-                        iface.IsGenericType
-                        && iface.OriginalDefinition.ToDisplayString(
-                            SymbolDisplayFormat.FullyQualifiedFormat
-                        )
-                            == "global::Microsoft.EntityFrameworkCore.IEntityTypeConfiguration<TEntity>"
-                    )
-                    {
-                        var entityFqn = iface
-                            .TypeArguments[0]
-                            .ToDisplayString(SymbolDisplayFormat.FullyQualifiedFormat);
-                        entityConfigs.Add(
-                            new EntityConfigInfo
-                            {
-                                ConfigFqn = typeSymbol.ToDisplayString(
-                                    SymbolDisplayFormat.FullyQualifiedFormat
-                                ),
-                                EntityFqn = entityFqn,
-                                ModuleName = moduleName,
-                                Location = SymbolHelpers.GetSourceLocation(typeSymbol),
-                            }
-                        );
-                        break;
-                    }
-                }
-            }
-        }
-    }
-
     private static void ScanContractInterfaces(
         INamespaceSymbol namespaceSymbol,
         string assemblyName,
@@ -1037,7 +837,9 @@ internal static class SymbolDiscovery
                                 ModuleName = moduleName,
                                 IsPublic = typeSymbol.DeclaredAccessibility == Accessibility.Public,
                                 IsAbstract = typeSymbol.IsAbstract,
-                                DependsOnDbContext = HasDbContextConstructorParam(typeSymbol),
+                                DependsOnDbContext = DbContextFinder.HasDbContextConstructorParam(
+                                    typeSymbol
+                                ),
                                 Location = SymbolHelpers.GetSourceLocation(typeSymbol),
                                 Lifetime = GetContractLifetime(typeSymbol),
                             }
